+++ tests/RocketeerTest.php
@@ -11,35 +11,35 @@ class RocketeerTest extends RocketeerTestCase
 
 	public function testCanGetAvailableConnections()
 	{
-		$connections = $this->app['rocketeer.rocketeer']->getAvailableConnections();
+		$connections = $this->app['rocketeer.connections']->getAvailableConnections();
 		$this->assertEquals(array('production', 'staging'), array_keys($connections));
 
 		$this->app['rocketeer.server']->setValue('connections.custom.username', 'foobar');
-		$connections = $this->app['rocketeer.rocketeer']->getAvailableConnections();
+		$connections = $this->app['rocketeer.connections']->getAvailableConnections();
 		$this->assertEquals(array('production', 'staging', 'custom'), array_keys($connections));
 	}
 
 	public function testCanGetCurrentConnection()
 	{
 		$this->swapConfig(array('rocketeer::default' => 'foobar'));
-		$this->assertEquals('production', $this->app['rocketeer.rocketeer']->getConnection());
+		$this->assertEquals('production', $this->app['rocketeer.connections']->getConnection());
 
 		$this->swapConfig(array('rocketeer::default' => 'production'));
-		$this->assertEquals('production', $this->app['rocketeer.rocketeer']->getConnection());
+		$this->assertEquals('production', $this->app['rocketeer.connections']->getConnection());
 
 		$this->swapConfig(array('rocketeer::default' => 'staging'));
-		$this->assertEquals('staging', $this->app['rocketeer.rocketeer']->getConnection());
+		$this->assertEquals('staging', $this->app['rocketeer.connections']->getConnection());
 	}
 
 	public function testCanChangeConnection()
 	{
-		$this->assertEquals('production', $this->app['rocketeer.rocketeer']->getConnection());
+		$this->assertEquals('production', $this->app['rocketeer.connections']->getConnection());
 
-		$this->app['rocketeer.rocketeer']->setConnection('staging');
-		$this->assertEquals('staging', $this->app['rocketeer.rocketeer']->getConnection());
+		$this->app['rocketeer.connections']->setConnection('staging');
+		$this->assertEquals('staging', $this->app['rocketeer.connections']->getConnection());
 
-		$this->app['rocketeer.rocketeer']->setConnections('staging,production');
-		$this->assertEquals(array('staging', 'production'), $this->app['rocketeer.rocketeer']->getConnections());
+		$this->app['rocketeer.connections']->setConnections('staging,production');
+		$this->assertEquals(array('staging', 'production'), $this->app['rocketeer.connections']->getConnections());
 	}
 
 	public function testCanUseSshRepository()
@@ -47,53 +47,53 @@ class RocketeerTest extends RocketeerTestCase
 		$repository = 'git@github.com:'.$this->repository;
 		$this->expectRepositoryConfig($repository, '', '');
 
-		$this->assertEquals($repository, $this->app['rocketeer.rocketeer']->getRepository());
+		$this->assertEquals($repository, $this->app['rocketeer.connections']->getRepository());
 	}
 
 	public function testCanUseHttpsRepository()
 	{
 		$this->expectRepositoryConfig('https://github.com/'.$this->repository, 'foobar', 'bar');
 
-		$this->assertEquals('https://foobar:bar@github.com/'.$this->repository, $this->app['rocketeer.rocketeer']->getRepository());
+		$this->assertEquals('https://foobar:bar@github.com/'.$this->repository, $this->app['rocketeer.connections']->getRepository());
 	}
 
 	public function testCanUseHttpsRepositoryWithUsernameProvided()
 	{
 		$this->expectRepositoryConfig('https://foobar@github.com/'.$this->repository, 'foobar', 'bar');
 
-		$this->assertEquals('https://foobar:bar@github.com/'.$this->repository, $this->app['rocketeer.rocketeer']->getRepository());
+		$this->assertEquals('https://foobar:bar@github.com/'.$this->repository, $this->app['rocketeer.connections']->getRepository());
 	}
 
 	public function testCanUseHttpsRepositoryWithOnlyUsernameProvided()
 	{
 		$this->expectRepositoryConfig('https://foobar@github.com/'.$this->repository, 'foobar', '');
 
-		$this->assertEquals('https://foobar@github.com/'.$this->repository, $this->app['rocketeer.rocketeer']->getRepository());
+		$this->assertEquals('https://foobar@github.com/'.$this->repository, $this->app['rocketeer.connections']->getRepository());
 	}
 
 	public function testCanCleanupProvidedRepositoryFromCredentials()
 	{
 		$this->expectRepositoryConfig('https://foobar@github.com/'.$this->repository, 'Anahkiasen', '');
 
-		$this->assertEquals('https://Anahkiasen@github.com/'.$this->repository, $this->app['rocketeer.rocketeer']->getRepository());
+		$this->assertEquals('https://Anahkiasen@github.com/'.$this->repository, $this->app['rocketeer.connections']->getRepository());
 	}
 
 	public function testCanUseHttpsRepositoryWithoutCredentials()
 	{
 		$this->expectRepositoryConfig('https://github.com/'.$this->repository, '', '');
 
-		$this->assertEquals('https://github.com/'.$this->repository, $this->app['rocketeer.rocketeer']->getRepository());
+		$this->assertEquals('https://github.com/'.$this->repository, $this->app['rocketeer.connections']->getRepository());
 	}
 
 	public function testCanCheckIfRepositoryNeedsCredentials()
 	{
 		$this->expectRepositoryConfig('https://github.com/'.$this->repository, '', '');
-		$this->assertTrue($this->app['rocketeer.rocketeer']->needsCredentials());
+		$this->assertTrue($this->app['rocketeer.connections']->needsCredentials());
 	}
 
 	public function testCangetRepositoryBranch()
 	{
-		$this->assertEquals('master', $this->app['rocketeer.rocketeer']->getRepositoryBranch());
+		$this->assertEquals('master', $this->app['rocketeer.connections']->getRepositoryBranch());
 	}
 
 	public function testCanGetApplicationName()
@@ -108,7 +108,7 @@ class RocketeerTest extends RocketeerTestCase
 
 	public function testCanGetFolderWithStage()
 	{
-		$this->app['rocketeer.rocketeer']->setStage('test');
+		$this->app['rocketeer.connections']->setStage('test');
 
 		$this->assertEquals($this->server.'/test/current', $this->app['rocketeer.rocketeer']->getFolder('current'));
 	}
@@ -140,7 +140,7 @@ class RocketeerTest extends RocketeerTestCase
 		));
 
 		$this->assertEquals('master', $this->app['rocketeer.rocketeer']->getOption('scm.branch'));
-		$this->app['rocketeer.rocketeer']->setStage('staging');
+		$this->app['rocketeer.connections']->setStage('staging');
 		$this->assertEquals('staging', $this->app['rocketeer.rocketeer']->getOption('scm.branch'));
 	}
 
@@ -178,7 +178,7 @@ class RocketeerTest extends RocketeerTestCase
 
 	public function testFillsConnectionCredentialsHoles()
 	{
-		$connections = $this->app['rocketeer.rocketeer']->getAvailableConnections();
+		$connections = $this->app['rocketeer.connections']->getAvailableConnections();
 		$this->assertArrayHasKey('production', $connections);
 
 		$this->app['rocketeer.server']->setValue('connections', array(
@@ -191,7 +191,7 @@ class RocketeerTest extends RocketeerTestCase
 				'agent'     => ''
 			),
 		));
-		$connections = $this->app['rocketeer.rocketeer']->getAvailableConnections();
+		$connections = $this->app['rocketeer.connections']->getAvailableConnections();
 		$this->assertArrayHasKey('production', $connections);
 	}